pyproject.toml
[project]
name = "computer_use"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.0",
    "anthropic>=0.71.0",
    "asyncpg>=0.30.0",
    "fastapi[all]>=0.118.3",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "sqlalchemy[asyncio]>=2.0.44",
]
